third_party/blink/web_tests/fast/events/resize-subframe-expected.txt

CONSOLE MESSAGE: line 8: PASS
